review-service/requirements.txt
aiokafka==0.10.0
fastapi==0.110.0
uvicorn[standard]==0.30.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.7.1
python-dotenv==1.0.1
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
prometheus-fastapi-instrumentator==6.1.0